requirements.txt
alabaster					0.7.12
altgraph					0.17.4
argon2-cffi					21.3.0
argon2-cffi-bindings					21.2.0
arrow					1.2.2
astroid					2.12.12
asttokens					2.2.1
attrs					22.1.0
autopep8					1.6.0
Babel					2.10.3
backcall					0.2.0
beautifulsoup4					4.11.1
binaryornot					0.4.4
black					22.10.0
bleach					5.0.1
cachetools					5.3.2
certifi					2022.6.15
cffi					1.15.1
chardet					5.0.0
charset-normalizer					2.1.1
cloudpickle					2.1.0
colorama					0.4.5
coloredlogs					15.0.1
contourpy					1.0.6
cookiecutter					2.1.1
cryptography					37.0.4
cycler					0.11.0
Cython					0.29.32
debugpy					1.6.3
decorator					4.4.2
defusedxml					0.7.1
diff-match-patch					20200713
dill					0.3.5.1
docstring-to-markdown					0.10
docutils					0.19
entrypoints					0.4
et-xmlfile					1.1.0
executing					1.2.0
fastjsonschema					2.16.1
flake8					4.0.1
fonttools					4.37.1
humanfriendly					10.0
idna					3.3
imageio-ffmpeg					0.4.7
imagesize					1.4.1
importlib-metadata					4.12.0
importlib-resources					5.10.0
inflection					0.5.1
ipykernel					6.15.2
ipython					7.34.0
ipython-genutils					0.2.0
ipywidgets					8.0.2
isort					5.10.1
jaraco.classes					3.2.3
jedi					0.18.1
jellyfish					0.9.0
Jinja2					3.1.2
jinja2-time					0.2.0
jsonschema					4.14.0
jupyter					1.0.0
jupyter-console					6.4.4
jupyter-core					4.11.1
jupyter_client					7.3.5
jupyterlab-pygments					0.2.2
jupyterlab-widgets					3.0.3
keyboard					0.13.5
keyring					23.8.2
kiwisolver					1.4.4
lazy-object-proxy					1.7.1
light-process					0.0.7
lxml					4.9.1
MarkupSafe					2.1.1
matplotlib					3.5.3
matplotlib-inline					0.1.6
mccabe					0.6.1
mistune					2.0.4
more-itertools					9.0.0
moviepy					1.0.3
mpmath					1.2.1
mypy-extensions					0.4.3
nbclient					0.6.7
nbconvert					7.0.0
nbformat					5.4.0
nest-asyncio					1.5.5
notebook					6.4.12
numpy					1.22.4
numpydoc					1.4.0
opencv-contrib-python					4.6.0.66
openpyxl					3.0.10
package-parser					1.0.2
packaging					23.2
pandas					1.4.3
pandocfilters					1.5.0
paramiko					2.11.0
parso					0.8.3
pathspec					0.10.1
pefile					2023.2.7
pexpect					4.8.0
pickleshare					0.7.5
Pillow					9.3.0
pip					23.2.1
pkgutil_resolve_name					1.3.10
platformdirs					2.5.2
pluggy					1.0.0
proglog					0.1.10
prometheus-client					0.14.1
prompt-toolkit					3.0.30
psutil					5.9.1
ptyprocess					0.7.0
pure-eval					0.2.2
py2exe					0.13.0.1
pycodestyle					2.8.0
pycparser					2.21
pydocstyle					6.1.1
pyflakes					2.4.0
pygame					2.1.2
Pygments					2.13.0
pyinstaller					6.2.0
pyinstaller-hooks-contrib					2023.10
pylibimp					1.0.1
pylibimport					1.9.2
pylint					2.15.0
pylint-venv					2.3.0
pyls-spyder					0.4.0
pypiwin32					223
pypylon					1.8.0
pypylon-opencv-viewer					1.1.0
PyQt5					5.15.7
PyQt5-Qt5					5.15.2
PyQt5-sip					12.11.0
PyQtWebEngine					5.15.6
PyQtWebEngine-Qt5					5.15.2
pyreadline3					3.4.1
pyrsistent					0.18.1
PySimpleGUI					4.60.4
PySimpleGUIWeb					0.39.0
python-dateutil					2.8.2
python-lsp-black					1.2.1
python-lsp-jsonrpc					1.0.0
python-lsp-server					1.5.0
pytoolconfig					1.2.2
pytz					2022.2.1
pywin32					305
pywin32-ctypes					0.2.2
pywinpty					2.0.7
pyxargs					2.0.0
pyxdg					0.28
pyzmq					23.2.1
QDarkStyle					3.0.3
qstylizer					0.2.1
QtAwesome					1.1.1
qtconsole					5.3.2
QtPy					2.2.0
remi					2020.3.10
requests					2.28.1
rope					1.3.0
RuNAs					0.1
scikit-image					0.19.3
scipy					1.9.1
Send2Trash					1.8.0
setuptools					65.3.0
six					1.16.0
snowballstemmer					2.2.0
sortedcontainers					2.4.0
soupsieve					2.3.2.post1
Sphinx					5.1.1
sphinxcontrib-applehelp					1.0.2
sphinxcontrib-devhelp					1.0.2
sphinxcontrib-htmlhelp					2.0.0
sphinxcontrib-jsmath					1.0.1
sphinxcontrib-qthelp					1.0.3
sphinxcontrib-serializinghtml					1.1.5
spyder-kernels					2.3.3
spyder-terminal					1.2.2
stack-data					0.5.1
swig					4.0.2
sympy					1.11.1
terminado					0.15.0
text-unidecode					1.3
textdistance					4.4.0
threadpoolctl					3.1.0
three-merge					0.1.1
tinycss2					1.1.1
tk					0.1.0
toml					0.10.2
tomli					2.0.1
tomlkit					0.11.4
tornado					6.2
tqdm					4.65.0
traitlets					5.3.0
typing_extensions					4.3.0
ujson					5.4.0
Unidecode					1.3.7
UNKNOWN					0.0.0
urllib3					1.26.12
watchdog					2.1.9
wcwidth					0.2.5
webencodings					0.5.1
whatthepatch					1.0.2
wheel					0.37.1
widgetsnbextension					4.0.3
win10toast					0.9
wincertstore					0.2
wrapt					1.14.1
XlsxWriter					3.0.3
yapf					0.32.0
zipp					3.10.0
